component: abi_stability
