+++ .evergreen/generated_configs/variants.yml
@@ -1,4 +1,24 @@
 buildvariants:
+  - name: abi-stability-polyfill
+    display_name: ABI Stability Checks (polyfill)
+    display_tasks:
+      - name: ABI Stability Checks (polyfill)
+        execution_tasks:
+          - .abi-stability
+    expansions:
+      cxx_standard: "11"
+    tasks:
+      - name: tg-abi-stability
+  - name: abi-stability-stdlib
+    display_name: ABI Stability Checks (stdlib)
+    display_tasks:
+      - name: ABI Stability Checks (stdlib)
+        execution_tasks:
+          - .abi-stability
+    expansions:
+      cxx_standard: "17"
+    tasks:
+      - name: tg-abi-stability
   - name: stub
     display_name: stub
     tasks:
