+++ .mci.yml
@@ -27,22 +27,6 @@ post:
 #######################################
 
 tasks:
-    - name: abi-compliance-check
-      tags: ["abi-stability", "abi-compliance-check"]
-      run_on: "ubuntu2204-large"
-      commands:
-        - func: "abi-compliance-check"
-    - name: abidiff
-      tags: ["abi-stability", "abidiff"]
-      run_on: "ubuntu2204-large"
-      commands:
-        - func: "abidiff"
-    - name: abi-prohibited-symbols
-      tags: ["abi-stability", "abi-prohibited-symbols"]
-      run_on: "ubuntu2204-large"
-      commands:
-        - func: "abi-prohibited-symbols"
-
     - name: lint
       run_on: ubuntu1804-large
       commands:
@@ -549,27 +533,6 @@ tasks:
         - func: "upload augmented sbom"
 
 task_groups:
-  - name: tg-abi-stability
-    max_hosts: -1
-    setup_task_can_fail_task: true
-    setup_task:
-      - command: git.get_project
-        params:
-          directory: "mongo-cxx-driver"
-      - func: install_c_driver
-      - command: subprocess.exec
-        params:
-          binary: bash
-          args: [-c, mongo-cxx-driver/.evergreen-old/abi-stability-setup.sh]
-          include_expansions_in_env: [cxx_standard]
-    tasks: [".abi-stability"]
-    teardown_task_can_fail_task: true
-    teardown_task:
-      - command: subprocess.exec
-        params:
-          binary: bash
-          args: [-c, rm -rf *]
-
   - name: test_atlas_task_group_search_indexes_7.0
     setup_group:
       - func: "setup"
@@ -632,29 +595,6 @@ task_groups:
 #######################################
 
 buildvariants:
-    ##########################
-    #  ABI Stability Checks  #
-    ##########################
-    - name: abi-stability-polyfill
-      display_name: "ABI Stability Checks (polyfill)"
-      expansions:
-        cxx_standard: 11 # Use a polyfill library.
-      tasks:
-        - "tg-abi-stability"
-      display_tasks:
-        - name: "ABI Stability Checks (polyfill)"
-          execution_tasks: [".abi-stability"]
-
-    - name: abi-stability-stdlib
-      display_name: "ABI Stability Checks (stdlib)"
-      expansions:
-        cxx_standard: 17 # Use the stdlib.
-      tasks:
-        - "tg-abi-stability"
-      display_tasks:
-        - name: "ABI Stability Checks (stdlib)"
-          execution_tasks: [".abi-stability"]
-
     #######################################
     #  Standard MongoDB Integration Tests #
     #######################################
